Add files via upload
Hey Guys, This is my first GITHUB PROJECT. So let me tell you some information about this Project. The name of the project is Hangman. Hangman is a word thinking game in which you have to think the word, you have got 7 lives to guess the words there are about 2000 words that can be guessed. So, use your knowledge and save the stickman, or else he will be hanged to DEATH. This project is made using Pycharm 2018.3.7 version and Python 3.8(32 bit). You can also use VScode for making thios project.